Change MaxPageLimit constant type to uint16
The MaxPageLimit constant is intended to be used for setting up:
RequestsPaginator.Limit and WorkersPaginator.Limit.

Both of these fields are uint16 type, so the constant should also
be defined as that type.

+++ lists.go
@@ -47,7 +47,7 @@ const (
 	DirectionBackward ListDirection = true
 
 	// MaxPageLimit denotes maximum value that pageInfo.Limit can be.
-	MaxPageLimit = math.MaxUint16
+	MaxPageLimit = uint16(math.MaxUint16)
 )
 
 // SortInfo contains information needed to sort collections in Boruta (requests, workers).
